Generate the AI configuration endpoints like every other call
The three AI configuration calls sat outside the generated API: they were
hand-written #[napi] functions in but-napi, so they never reached the
#[but_api] name table, and every client had to hand-wire a forward.

- moved get/update/reset_ai_configuration into but-api as #[but_api(napi)],
  so the macro generates their wrappers and registers them in apiParamNames
- streaming stays in but-napi: it takes a callback, which the macro does not
  carry across the boundary
- they now declare cache effects like any other endpoint: the read provides
  the new AiConfiguration tag, the two writes invalidate it
- lite's main process drops its three hand-written forwards, so its overrides
  are purely electron capabilities again and the endpoints derive like the
  rest
- the TypeScript union types survive the move, and the field docs now reach
  the SDK
