+++ apps/lite/electron/src/ipc.ts
@@ -55,17 +55,14 @@ export const localEndpoints = [
 	"askpassSubmitPromptResponse",
 	"clipboardWriteText",
 	"fullScreenChange",
-	"getAiConfiguration",
 	"getVersion",
 	"isFullScreen",
 	"openInWebBrowser",
 	"pathJoin",
 	"pickDirectory",
 	"readGUISettings",
-	"resetAiConfiguration",
 	"showNativeMenu",
 	"streamAiResponse",
-	"updateAiConfiguration",
 	"watcherStopAll",
 	"watcherSubscribe",
 	"watcherUnsubscribe",
