+++ crates/but-napi/Cargo.toml
@@ -41,7 +41,6 @@ but-core.workspace = true
 but-ctx.workspace = true
 but-llm.workspace = true
 but-path.workspace = true
-but-secret.workspace = true
 but-settings.workspace = true
 gix.workspace = true
 gitbutler-watcher.workspace = true
